resolver: inline small h2 module
